requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
greenlet==3.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.24.4
pandas==1.5.1
PyMySQL==1.0.2
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
pytz==2024.2
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==3.0.4
zipp==3.20.2
